Shout fix all requested changes to PR.
